task: Split arweave into public and private
