+++ src/arweave.ts
@@ -1,16 +1,9 @@
 import * as fs from 'fs';
 import { JWKInterface } from 'arweave/node/lib/wallet';
-import { appName, appVersion, arFSVersion, weightedRandom } from './common';
-import { ArDriveUser, ArFSDriveMetaData, ArFSFileMetaData } from './types/base_Types';
-import {
-	ArFSDriveEntity,
-	ArFSFileData,
-	ArFSFileFolderEntity,
-	ArFSPrivateDriveEntity,
-	ArFSPrivateFileData,
-	ArFSPrivateFileFolderEntity,
-	Wallet
-} from './types/arfs_Types';
+import { weightedRandom } from './common';
+import { appName, appVersion, arFSVersion } from './constants';
+
+import { ArFSDriveEntity, ArFSFileData, ArFSFileFolderEntity, Wallet } from './types/arfs_Types';
 import { readContract } from 'smartweave';
 import Arweave from 'arweave';
 import deepHash from 'arweave/node/lib/deepHash';
@@ -155,42 +148,6 @@ export async function createDriveTransaction(
 	return transaction;
 }
 
-// Creates an arweave transaction to upload a drive entity
-export async function createPrivateDriveTransaction(
-	driveJSON: Buffer, // must be an encrypted buffer
-	driveMetaData: ArFSPrivateDriveEntity,
-	walletPrivateKey?: JWKInterface
-): Promise<Transaction> {
-	// Create transaction
-	let transaction: Transaction;
-	if (walletPrivateKey) {
-		transaction = await arweave.createTransaction({ data: driveJSON }, walletPrivateKey);
-	} else {
-		transaction = await arweave.createTransaction({ data: driveJSON }); // Will use ArConnect if no wallet present
-	}
-	// Tag file with ArFS Tags
-	transaction.addTag('App-Name', driveMetaData.appName);
-	transaction.addTag('App-Version', driveMetaData.appVersion);
-	transaction.addTag('Unix-Time', driveMetaData.unixTime.toString());
-	transaction.addTag('Drive-Id', driveMetaData.driveId);
-	transaction.addTag('Drive-Privacy', driveMetaData.drivePrivacy);
-	transaction.addTag('Content-Type', driveMetaData.contentType);
-	// Tag file with Content-Type, Cipher and Cipher-IV and Drive-Auth-Mode
-	transaction.addTag('Cipher', driveMetaData.cipher);
-	transaction.addTag('Cipher-IV', driveMetaData.cipherIV);
-	transaction.addTag('Drive-Auth-Mode', driveMetaData.driveAuthMode);
-	transaction.addTag('ArFS', driveMetaData.arFS);
-	transaction.addTag('Entity-Type', 'drive');
-
-	// Sign file
-	if (walletPrivateKey) {
-		await arweave.transactions.sign(transaction, walletPrivateKey);
-	} else {
-		await arweave.transactions.sign(transaction); // Will use ArConnect if no wallet present
-	}
-	return transaction;
-}
-
 // This will prepare and sign v2 data transaction using ArFS File Data Tags
 export async function createFileDataTransaction(
 	fileData: Buffer,
@@ -209,37 +166,6 @@ export async function createFileDataTransaction(
 	transaction.addTag('App-Name', fileMetaData.appName);
 	transaction.addTag('App-Version', fileMetaData.appVersion);
 	transaction.addTag('Content-Type', fileMetaData.contentType);
-
-	// Sign the transaction
-	if (walletPrivateKey) {
-		await arweave.transactions.sign(transaction, walletPrivateKey);
-	} else {
-		await arweave.transactions.sign(transaction); // Will use ArConnect if no wallet present
-	}
-
-	return transaction;
-}
-
-// This will prepare and sign a private v2 data transaction using ArFS File Data Tags including privacy tags
-export async function createPrivateFileDataTransaction(
-	fileData: Buffer, // the buffer must already be encrypted
-	fileMetaData: ArFSPrivateFileData,
-	walletPrivateKey?: JWKInterface
-): Promise<Transaction> {
-	let transaction: Transaction;
-	// Create the arweave transaction using the file data and private key
-	if (walletPrivateKey) {
-		transaction = await arweave.createTransaction({ data: fileData }, walletPrivateKey);
-	} else {
-		transaction = await arweave.createTransaction({ data: fileData }); // Will use ArConnect if no wallet present
-	}
-
-	// Tag file with Content-Type, Cipher and Cipher-IV
-	transaction.addTag('App-Name', fileMetaData.appName);
-	transaction.addTag('App-Version', fileMetaData.appVersion);
-	transaction.addTag('Content-Type', 'application/octet-stream');
-	transaction.addTag('Cipher', fileMetaData.cipher);
-	transaction.addTag('Cipher-IV', fileMetaData.cipherIV);
 
 	// Sign the transaction
 	if (walletPrivateKey) {
@@ -296,52 +222,6 @@ export async function createFileFolderMetaDataTransaction(
 	return transaction;
 }
 
-// This will prepare and sign a private v2 data transaction using ArFS File Metadata Tags including privacy tags
-export async function createPrivateFileFolderMetaDataTransaction(
-	metaData: ArFSPrivateFileFolderEntity,
-	secondaryFileMetaData: Buffer, // the buffer must already be encrypted
-	walletPrivateKey?: JWKInterface
-): Promise<Transaction> {
-	let transaction: Transaction;
-	if (walletPrivateKey) {
-		// Create the arweave transaction using the file data and private key
-		transaction = await arweave.createTransaction({ data: secondaryFileMetaData }, walletPrivateKey);
-	} else {
-		transaction = await arweave.createTransaction({ data: secondaryFileMetaData }); // Will use ArConnect if no wallet present
-	}
-
-	// Tag file with ArFS Tags including tags needed for privacy
-	transaction.addTag('App-Name', metaData.appName);
-	transaction.addTag('App-Version', metaData.appVersion);
-	transaction.addTag('Unix-Time', metaData.unixTime.toString());
-	transaction.addTag('Content-Type', 'application/octet-stream');
-	transaction.addTag('Cipher', metaData.cipher);
-	transaction.addTag('Cipher-IV', metaData.cipherIV);
-	transaction.addTag('ArFS', metaData.arFS);
-	transaction.addTag('Entity-Type', metaData.entityType);
-	transaction.addTag('Drive-Id', metaData.driveId);
-
-	// Add file or folder specific tags
-	if (metaData.entityType === 'file') {
-		transaction.addTag('File-Id', metaData.entityId);
-		transaction.addTag('Parent-Folder-Id', metaData.parentFolderId);
-	} else {
-		transaction.addTag('Folder-Id', metaData.entityId);
-		if (metaData.parentFolderId !== '0') {
-			transaction.addTag('Parent-Folder-Id', metaData.parentFolderId);
-		}
-	}
-
-	// Sign the transaction
-	if (walletPrivateKey) {
-		await arweave.transactions.sign(transaction, walletPrivateKey);
-	} else {
-		await arweave.transactions.sign(transaction); // Will use ArConnect if no wallet present
-	}
-
-	return transaction;
-}
-
 // Creates an arweave data item transaction (ANS-102) using ArFS Tags
 export async function createFileDataItemTransaction(
 	fileData: Buffer,
@@ -357,35 +237,6 @@ export async function createFileDataItemTransaction(
 		arBundles.addTag(item, 'App-Version', fileMetaData.appVersion);
 		// Only tag the file with public tags
 		arBundles.addTag(item, 'Content-Type', fileMetaData.contentType);
-
-		// Sign the data, ready to be added to a bundle
-		const signedItem = await arBundles.sign(item, walletPrivateKey);
-		return signedItem;
-	} catch (err) {
-		console.log('Error creating data item');
-		console.log(err);
-		return 'Error';
-	}
-}
-
-// Creates a private arweave data item transaction (ANS-102) using ArFS Tags including privacy tags
-export async function createPrivateFileDataItemTransaction(
-	fileData: Buffer, // the buffer must already be encrypted
-	fileMetaData: ArFSPrivateFileData,
-	walletPrivateKey: JWKInterface
-): Promise<DataItemJson | string> {
-	try {
-		// Create the item using the data buffer
-		const item = await arBundles.createData({ data: fileData }, walletPrivateKey);
-
-		// Tag file with common tags
-		arBundles.addTag(item, 'App-Name', fileMetaData.appName);
-		arBundles.addTag(item, 'App-Version', fileMetaData.appVersion);
-
-		// Tag file with Privacy tags, Content-Type, Cipher and Cipher-IV
-		arBundles.addTag(item, 'Content-Type', 'application/octet-stream');
-		arBundles.addTag(item, 'Cipher', fileMetaData.cipher);
-		arBundles.addTag(item, 'Cipher-IV', fileMetaData.cipherIV);
 
 		// Sign the data, ready to be added to a bundle
 		const signedItem = await arBundles.sign(item, walletPrivateKey);
@@ -439,52 +290,6 @@ export async function createFileFolderMetaDataItemTransaction(
 	}
 }
 
-// Creates an arweave data item transaction (ANS-102) using ArFS Tags
-export async function createPrivateFileFolderMetaDataItemTransaction(
-	metaData: ArFSPrivateFileFolderEntity,
-	secondaryFileMetaData: Buffer, // the buffer must already be encrypted
-	walletPrivateKey: JWKInterface
-): Promise<DataItemJson | string> {
-	try {
-		// Create the item using the data buffer or string
-		const item = await arBundles.createData({ data: secondaryFileMetaData }, walletPrivateKey);
-
-		// Tag file
-		arBundles.addTag(item, 'App-Name', metaData.appName);
-		arBundles.addTag(item, 'App-Version', metaData.appVersion);
-		arBundles.addTag(item, 'Unix-Time', metaData.unixTime.toString());
-		// If the file is private, we use extra tags
-		// Tag file with Content-Type, Cipher and Cipher-IV
-		arBundles.addTag(item, 'Content-Type', 'application/octet-stream');
-		arBundles.addTag(item, 'Cipher', metaData.cipher);
-		arBundles.addTag(item, 'Cipher-IV', metaData.cipherIV);
-		arBundles.addTag(item, 'ArFS', arFSVersion);
-		arBundles.addTag(item, 'Entity-Type', metaData.entityType);
-		arBundles.addTag(item, 'Drive-Id', metaData.driveId);
-		arBundles.addTag(item, 'File-Id', metaData.entityId);
-
-		// Add file or folder specific tags
-		if (metaData.entityType === 'file') {
-			arBundles.addTag(item, 'File-Id', metaData.entityId);
-			arBundles.addTag(item, 'Parent-Folder-Id', metaData.parentFolderId);
-		} else {
-			arBundles.addTag(item, 'Folder-Id', metaData.entityId);
-			if (metaData.parentFolderId !== '0') {
-				// If the parentFolderId is 0, then this is a root folder
-				arBundles.addTag(item, 'Parent-Folder-Id', metaData.parentFolderId);
-			}
-		}
-
-		// Sign the data, ready to be added to a bundle
-		const signedItem = await arBundles.sign(item, walletPrivateKey);
-		return signedItem;
-	} catch (err) {
-		console.log('Error creating data item');
-		console.log(err);
-		return 'Error';
-	}
-}
-
 // Creates a bundled data transaction (ANS-102)
 export async function createBundledDataTransaction(
 	walletPrivateKey: JWKInterface,
@@ -507,234 +312,6 @@ export async function createBundledDataTransaction(
 
 		// Sign the bundle
 		await arweave.transactions.sign(transaction, walletPrivateKey);
-		return transaction;
-	} catch (err) {
-		console.log('Error creating data bundle');
-		console.log(err);
-		return null;
-	}
-}
-
-// Creates an arweave transaction to upload encrypted private ardrive metadata
-// SPLIT INTO createPrivateDriveTransaction and createDriveTransaction
-export async function prepareArFSDriveTransaction(
-	user: ArDriveUser,
-	driveJSON: string | Buffer,
-	driveMetaData: ArFSDriveMetaData
-): Promise<Transaction> {
-	// Create transaction
-	const transaction = await arweave.createTransaction({ data: driveJSON }, JSON.parse(user.walletPrivateKey));
-
-	// Tag file with ArFS Tags
-	transaction.addTag('App-Name', appName);
-	transaction.addTag('App-Version', appVersion);
-	transaction.addTag('Unix-Time', driveMetaData.unixTime.toString());
-	transaction.addTag('Drive-Id', driveMetaData.driveId);
-	transaction.addTag('Drive-Privacy', driveMetaData.drivePrivacy);
-	if (driveMetaData.drivePrivacy === 'private') {
-		// If the file is private, we use extra tags
-		// Tag file with Content-Type, Cipher and Cipher-IV and Drive-Auth-Mode
-		transaction.addTag('Content-Type', 'application/octet-stream');
-		transaction.addTag('Cipher', driveMetaData.cipher);
-		transaction.addTag('Cipher-IV', driveMetaData.cipherIV);
-		transaction.addTag('Drive-Auth-Mode', driveMetaData.driveAuthMode);
-	} else {
-		// Tag file with public tags only
-		transaction.addTag('Content-Type', 'application/json');
-	}
-	transaction.addTag('ArFS', arFSVersion);
-	transaction.addTag('Entity-Type', 'drive');
-
-	// Sign file
-	await arweave.transactions.sign(transaction, JSON.parse(user.walletPrivateKey));
-	return transaction;
-}
-
-// This will prepare and sign v2 data transaction using ArFS File Data Tags
-// SPLIT INTO createPrivateFileDataTransaction and createFileDataTransaction
-export async function prepareArFSDataTransaction(
-	user: ArDriveUser,
-	fileData: Buffer,
-	fileMetaData: ArFSFileMetaData
-): Promise<Transaction> {
-	// Create the arweave transaction using the file data and private key
-	const transaction = await arweave.createTransaction({ data: fileData }, JSON.parse(user.walletPrivateKey));
-
-	// If the file is not public, we must encrypt it
-	if (fileMetaData.isPublic === 0) {
-		// Tag file with Content-Type, Cipher and Cipher-IV
-		transaction.addTag('App-Name', appName);
-		transaction.addTag('App-Version', appVersion);
-		transaction.addTag('Content-Type', 'application/octet-stream');
-		transaction.addTag('Cipher', fileMetaData.cipher);
-		transaction.addTag('Cipher-IV', fileMetaData.dataCipherIV);
-	} else {
-		// Tag file with public tags only
-		transaction.addTag('App-Name', appName);
-		transaction.addTag('App-Version', appVersion);
-		transaction.addTag('Content-Type', fileMetaData.contentType);
-	}
-
-	// Sign file
-	await arweave.transactions.sign(transaction, JSON.parse(user.walletPrivateKey));
-	return transaction;
-}
-
-// This will prepare and sign v2 data transaction using ArFS File Metadata Tags
-// SPLIT INTO createPrivateFileFolderMetaDataItemTransaction and createFileFolderMetaDataItemTransaction
-export async function prepareArFSMetaDataTransaction(
-	user: ArDriveUser,
-	fileMetaData: ArFSFileMetaData,
-	secondaryFileMetaData: string | Buffer
-): Promise<Transaction> {
-	// Create the arweave transaction using the file data and private key
-	const transaction = await arweave.createTransaction(
-		{ data: secondaryFileMetaData },
-		JSON.parse(user.walletPrivateKey)
-	);
-
-	// Tag file with ArFS Tags
-	transaction.addTag('App-Name', appName);
-	transaction.addTag('App-Version', appVersion);
-	transaction.addTag('Unix-Time', fileMetaData.unixTime.toString());
-	if (fileMetaData.isPublic === 0) {
-		// If the file is private, we use extra tags
-		// Tag file with Content-Type, Cipher and Cipher-IV
-		transaction.addTag('Content-Type', 'application/octet-stream');
-		transaction.addTag('Cipher', fileMetaData.cipher);
-		transaction.addTag('Cipher-IV', fileMetaData.metaDataCipherIV);
-	} else {
-		// Tag file with public tags only
-		transaction.addTag('Content-Type', fileMetaData.contentType);
-	}
-	transaction.addTag('ArFS', arFSVersion);
-	transaction.addTag('Entity-Type', fileMetaData.entityType);
-	transaction.addTag('Drive-Id', fileMetaData.driveId);
-
-	// Add file or folder specific tags
-	if (fileMetaData.entityType === 'file') {
-		transaction.addTag('File-Id', fileMetaData.fileId);
-		transaction.addTag('Parent-Folder-Id', fileMetaData.parentFolderId);
-	} else {
-		transaction.addTag('Folder-Id', fileMetaData.fileId);
-		if (fileMetaData.parentFolderId !== '0') {
-			// Root folder transactions do not have Parent-Folder-Id
-			transaction.addTag('Parent-Folder-Id', fileMetaData.parentFolderId);
-		}
-	}
-
-	// Sign transaction
-	await arweave.transactions.sign(transaction, JSON.parse(user.walletPrivateKey));
-	return transaction;
-}
-
-// Creates an arweave data item transaction (ANS-102) using ArFS Tags
-// SPLIT INTO createPrivateFileDataItemTransaction and createFileDataItemTransaction
-export async function prepareArFSDataItemTransaction(
-	user: ArDriveUser,
-	fileData: Buffer,
-	fileMetaData: ArFSFileMetaData
-): Promise<DataItemJson | null> {
-	try {
-		// Create the item using the data buffer
-		const item = await arBundles.createData({ data: fileData }, JSON.parse(user.walletPrivateKey));
-
-		// Tag file with common tags
-		arBundles.addTag(item, 'App-Name', appName);
-		arBundles.addTag(item, 'App-Version', appVersion);
-		if (fileMetaData.isPublic === 0) {
-			// If the file is private, we use extra tags
-			// Tag file with Privacy tags, Content-Type, Cipher and Cipher-IV
-			arBundles.addTag(item, 'Content-Type', 'application/octet-stream');
-			arBundles.addTag(item, 'Cipher', fileMetaData.cipher);
-			arBundles.addTag(item, 'Cipher-IV', fileMetaData.dataCipherIV);
-		} else {
-			// Only tag the file with public tags
-			arBundles.addTag(item, 'Content-Type', fileMetaData.contentType);
-		}
-
-		// Sign the data, ready to be added to a bundle
-		const signedItem = await arBundles.sign(item, JSON.parse(user.walletPrivateKey));
-		return signedItem;
-	} catch (err) {
-		console.log('Error creating data item');
-		console.log(err);
-		return null;
-	}
-}
-
-// Creates an arweave data item transaction (ANS-102) using ArFS Tags
-// SPLIT INTO createPrivateFileMetaDataItemTransaction and createFileMetaDataItemTransaction
-export async function prepareArFSMetaDataItemTransaction(
-	user: ArDriveUser,
-	fileMetaData: ArFSFileMetaData,
-	secondaryFileMetaData: string | Buffer
-): Promise<DataItemJson | null> {
-	try {
-		// Create the item using the data buffer or string
-		const item = await arBundles.createData({ data: secondaryFileMetaData }, JSON.parse(user.walletPrivateKey));
-
-		// Tag file
-		arBundles.addTag(item, 'App-Name', appName);
-		arBundles.addTag(item, 'App-Version', appVersion);
-		arBundles.addTag(item, 'Unix-Time', fileMetaData.unixTime.toString());
-		if (fileMetaData.isPublic === 0) {
-			// If the file is private, we use extra tags
-			// Tag file with Content-Type, Cipher and Cipher-IV
-			arBundles.addTag(item, 'Content-Type', 'application/octet-stream');
-			arBundles.addTag(item, 'Cipher', fileMetaData.cipher);
-			arBundles.addTag(item, 'Cipher-IV', fileMetaData.metaDataCipherIV);
-		} else {
-			arBundles.addTag(item, 'Content-Type', 'application/json');
-		}
-		arBundles.addTag(item, 'ArFS', arFSVersion);
-		arBundles.addTag(item, 'Entity-Type', fileMetaData.entityType);
-		arBundles.addTag(item, 'Drive-Id', fileMetaData.driveId);
-
-		// Add file or folder specific tags
-		if (fileMetaData.entityType === 'file') {
-			arBundles.addTag(item, 'File-Id', fileMetaData.fileId);
-			arBundles.addTag(item, 'Parent-Folder-Id', fileMetaData.parentFolderId);
-		} else {
-			arBundles.addTag(item, 'Folder-Id', fileMetaData.fileId);
-			if (fileMetaData.parentFolderId !== '0') {
-				arBundles.addTag(item, 'Parent-Folder-Id', fileMetaData.parentFolderId);
-			}
-		}
-
-		// Sign the data, ready to be added to a bundle
-		const signedItem = await arBundles.sign(item, JSON.parse(user.walletPrivateKey));
-		return signedItem;
-	} catch (err) {
-		console.log('Error creating data item');
-		console.log(err);
-		return null;
-	}
-}
-
-// Creates a bundled data transaction
-// MOVED TO createBundledDataTransaction
-export async function prepareArFSBundledDataTransaction(
-	user: ArDriveUser,
-	items: DataItemJson[]
-): Promise<Transaction | null> {
-	try {
-		// Bundle up all individual items into a single data bundle
-		const dataBundle = await arBundles.bundleData(items);
-		const dataBuffer: Buffer = Buffer.from(JSON.stringify(dataBundle));
-
-		// Create the transaction for the entire data bundle
-		const transaction = await arweave.createTransaction({ data: dataBuffer }, JSON.parse(user.walletPrivateKey));
-
-		// Tag file
-		transaction.addTag('App-Name', appName);
-		transaction.addTag('App-Version', appVersion);
-		transaction.addTag('Bundle-Format', 'json');
-		transaction.addTag('Bundle-Version', '1.0.0');
-		transaction.addTag('Content-Type', 'application/json');
-
-		// Sign the bundle
-		await arweave.transactions.sign(transaction, JSON.parse(user.walletPrivateKey));
 		return transaction;
 	} catch (err) {
 		console.log('Error creating data bundle');
